+++ src/lib/stores.js
@@ -1,3 +1,3 @@
 import { writable } from 'svelte/store';
 
-export const started = writable(false);
+export const score = writable(0);
