+++ src/routes/+layout.svelte
@@ -1,0 +1,7 @@
+<nav>
+	<a href="/">home</a>
+	<a href="/about">about</a>
+</nav>
+
+<slot />
+
